requirements.txt
numpy
matplotlib
plotly
scipy
Pillow